refactor: calendar - update material and fluent styles
