+++ src/components/calendar/themes/material/years-view.material.scss
@@ -1,6 +1,6 @@
-@import '../year-month';
+@use '../year-month' as *;
 
-:host {
+[part='years-view'] {
     @extend %host;
 }
