+++ src/components/calendar/themes/shared-calendar.scss
@@ -1,4 +1,4 @@
-@use '../../styles/utilities' as *;
+@use '../../../styles/utilities/index' as *;
 
 $date-size-large: rem(32px);
 $date-size-medium: rem(28px);
