+++ src/components/calendar/themes/year-month.scss
@@ -1,4 +1,4 @@
-@use '../../styles/utilities' as *;
+@use '../../../styles/utilities/index' as *;
 @use 'sass:math';
 
 %host {
