+++ src/styles/themes/dark/bootstrap.scss
@@ -11,7 +11,6 @@
 @use '../../../components/list/list.bootstrap' as list with (
     $separator-color: utils.color(gray, 100)
 );
-@use '../../../components/calendar/calendar.bootstrap' as calendar;
 @use '../../../components/input/input.bootstrap' as input with ($input-background: transparent);
 @use '../../../components/slider/slider.bootstrap' as slider with (
     $inactive-background: utils.color(secondary, 500, .48)
@@ -37,7 +36,6 @@ $palette: utils.palette(
 @include switch.theme();
 @include navbar.theme();
 @include list.theme();
-@include calendar.theme();
 @include input.theme();
 @include icon-button.theme();
 @include slider.theme();
