+++ src/styles/themes/light/fluent.scss
@@ -7,7 +7,6 @@
 @use '../../../components/radio/radio.fluent' as radio;
 @use '../../../components/list/list.fluent' as list;
 @use '../../../components/checkbox/switch.fluent' as switch;
-@use '../../../components/calendar/calendar.fluent' as calendar;
 @use '../../../components/input/input.fluent' as input;
 @use '../../../components/navbar/navbar.fluent' as navbar;
 @use '../../../components/slider/slider.fluent' as slider;
@@ -27,7 +26,6 @@ $palette: utils.palette(
 @include radio.theme();
 @include list.theme();
 @include switch.theme();
-@include calendar.theme();
 @include input.theme();
 @include icon-button.theme();
 @include navbar.theme();
